完成惹，完成惹，完成惹，完成惹，完成惹，完成惹

+++ ManagerClient/InfoForm.cs
@@ -12,6 +12,8 @@ using MaterialSkin;
 using MaterialSkin.Controls;
 using Sunny;
 using Sunny.UI;
+using System.IO;
+using System.Collections;
 
 namespace ManagerClient
 {
@@ -29,6 +31,9 @@ namespace ManagerClient
                            1 , 1 , 1 ,1 ,
                            1 , 1 , 1 ,1 ,
                            1 , 1 , 1 , 1};
+
+        public bool? Frame { get; private set; }
+
         public InfoForm()
         {
             InitializeComponent();
@@ -64,7 +69,12 @@ namespace ManagerClient
             String[] parm = { };
             datalist = conn.selelectWorkerInf(sql, parm);
             if (datalist.Count <= 0)
+            {
                 MessageBox.Show("没有数据");
+                this.uiDataGridView_worker.DataSource = datalist;
+                dataListSave = datalist.GetRange(0, datalist.Count); // 深度拷贝
+                this.uiDataGridView_worker.Refresh();
+            }
             else
             {
                 this.uiDataGridView_worker.DataSource = datalist;
@@ -154,7 +164,7 @@ namespace ManagerClient
                 DataLine.Append(strT + ",");
 
                 string FileName = DateTime.Now.Year.ToString() + "-" + DateTime.Now.Month.ToString() + "-" + DateTime.Now.Day.ToString();
-                string FilePath = DataFileRootPath + "\\" + FileName + "-report.txt";
+                string FilePath = DataFileRootPath + "\\" + FileName + "-report.csv";
                 System.IO.StreamWriter stream = new System.IO.StreamWriter(FilePath, false, Encoding.UTF8);
 
                 stream.WriteLine(DataLine);
@@ -182,14 +192,17 @@ namespace ManagerClient
         {
             int col = e.ColumnIndex; // 选中的列
             int cell = e.RowIndex; // 选中的行
-            string userId = uiDataGridView_worker.Rows[0].Cells[cell].Value.ToString();// 读取账号回头用来修改
+            string userId = uiDataGridView_worker.Rows[cell].Cells[0].Value.ToString();// 读取账号回头用来修改
             string changedStringName = "";
             string changedValue = "";
+            
             //MessageBox.Show("选中 列 " + col + "  选中 行 " + cell);
-            //MessageBox.Show(dataGridView_worker.Rows[cell].Cells[col].Value.ToString());
+            //MessageBox.Show(uiDataGridView_worker.Rows[cell].Cells[col].Value.ToString());
+
             changedValue = uiDataGridView_worker.Rows[cell].Cells[col].Value.ToString();
             changedStringName = this.uiDataGridView_worker.Columns[e.ColumnIndex].Name;
-            MessageBox.Show(changedStringName + " " + changedValue + " " + userId);
+            
+            //MessageBox.Show(changedStringName + " " + changedValue + " " + userId);
             WorkerService service = new WorkerService();
 
             int cnt = service.doWorkerInfoChangeByItem(changedStringName, changedValue, userId);
@@ -355,7 +368,72 @@ namespace ManagerClient
 
         private void uiButton_readInfo_Click(object sender, EventArgs e)
         {
+            OpenFileDialog saveFile = new OpenFileDialog();
+            saveFile.Filter = "所有Csv文件|*.csv";//设置文件类型
+            saveFile.Title = "导出数据";//设置标题
 
+            saveFile.AddExtension = true;//是否自动增加所辍名
+            saveFile.AutoUpgradeEnabled = true;//是否随系统升级而升级外观
+            if (saveFile.ShowDialog() == DialogResult.OK)//如果点的是确定就得到文件路径
+            {
+                String ImageFileName = saveFile.FileName;//得到文件路径
+                MessageBox.Show(ImageFileName);
+                try
+                {
+                    using(StreamReader sr = new StreamReader(ImageFileName))
+                    {
+                        String line;
+                        int cnt = 0;
+                        Hashtable setInf = new Hashtable();
+                        for(int i = 0; i < datalist.Count; i++)
+                        {
+                            setInf.Add(datalist[i].WorkerNumber,datalist[i]);
+                        }
+                        WorkerService service = new WorkerService();
+                        while((line = sr.ReadLine()) != null)
+                        {
+                            if(cnt > 1)
+                            {
+                                MessageBox.Show(line);
+                                String[] items = line.Split(',');
+                                if (setInf.ContainsKey(items[1]))
+                                {
+                                    continue;
+                                }
+                                else
+                                {
+                                    WorkerInf inf_new = new WorkerInf();
+                                    //inf_new.WorkerNumber = items;
+                                    inf_new.WorkerNumber = items[1];
+                                    inf_new.WorkerPassword = items[2];
+                                    inf_new.WorkerName = items[3];
+                                    inf_new.WorkerSex = int.Parse(items[4]);
+                                    inf_new.WorkerBirth = items[5];
+                                    inf_new.WorkerHomeAddr = items[6];
+                                    inf_new.WorkerEdu = int.Parse(items[7]);
+                                    inf_new.WorkerLv =int.Parse(items[8]);
+                                    inf_new.WorkerTime = int.Parse(items[9]);
+                                    inf_new.WorkerJoinTime = items[10];
+                                    inf_new.WorkerBaseRec = float.Parse(items[11]);
+                                    inf_new.WorkerAdd = items[12];
+                                    inf_new.WorkerTel = items[13];
+                                    service.insertWorkerInf(inf_new);
+                                }
+                            }
+                            cnt++;
+                        }
+                    }
+                    reRoad();
+                }
+                catch (Exception)
+                {
+
+                }
+                finally
+                {
+
+                }
+            }
         }
 
         private void uiButton_infoSave_Click(object sender, EventArgs e)
@@ -386,6 +464,10 @@ namespace ManagerClient
 
         private void uiButton_clearInfo_Click(object sender, EventArgs e)
         {
+            if (!ShowAskDialog("是否确定清空"))
+            {
+                return;
+            }
             String[] str = { };
             int lines = MysqlConnectorHelper.exectUID("delete from workertab;", str);
             MessageBox.Show("删除了: " + lines + " 条记录");
@@ -419,6 +501,37 @@ namespace ManagerClient
                 this.uiDataGridView_worker.Refresh();
                 dataListSave = datalist.GetRange(0, datalist.Count); // 深度拷贝
             }
+        }
+
+        private void uiDataGridView_worker_CellContentDoubleClick(object sender, DataGridViewCellEventArgs e)
+        {        
+            int col = e.ColumnIndex; // 选中的列
+            int cell = e.RowIndex; // 选中的行
+            string userId = uiDataGridView_worker.Rows[cell].Cells[0].Value.ToString();// 读取账号回头用来修改
+            string changedStringName = "";
+            string changedValue = "";
+
+            //MessageBox.Show("选中 列 " + col + "  选中 行 " + cell);
+            //MessageBox.Show(uiDataGridView_worker.Rows[cell].Cells[col].Value.ToString());
+
+            changedValue = uiDataGridView_worker.Rows[cell].Cells[col].Value.ToString();
+            changedStringName = this.uiDataGridView_worker.Columns[e.ColumnIndex].Name;
+
+            if (!ShowAskDialog("是否删除员工："+ userId))
+            {
+                return;
+            }
+
+            //MessageBox.Show(changedStringName + " " + changedValue + " " + userId);
+            WorkerService service = new WorkerService();
+            //MessageBox.Show(userId);
+            int cnt = service.doDeleteById(userId);
+            //int cnt = 2;
+            if (cnt <= 0)
+            {
+                MessageBox.Show("删除失败");
+            }
+            reRoad();
         }
     }
 }
